refactor(status-panel): split into core, cmd, and legacy files
Split the 473-line status-panel.go into three focused modules:
- status-panel-core.go: Core logic with client interface pattern
- status-panel-cmd.go: Cobra command definition and registration
- status-panel-legacy.go: Legacy testing infrastructure with global functions

Removes status-panel.go from lint exclusion list since none of the split
files exceed the 500-line threshold, improving code maintainability.
